web/requirements.txt
Django==3.0
psycopg2-binary==2.8.3
gunicorn==20.0.4
boto3==1.12.4
pillow==7.0.0